python/agents/travel-research/pyproject.toml
[project]
name = "travel-research"
version = "0.1.0"
description = "A multi-agent travel research system that produces PDF reports with visuals and itineraries"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
    "fpdf2>=2.8.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
example_question = "Research Tokyo for a 5-day trip"

[tool.agent-starter-pack.settings]
agent_directory = "travel_research"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
